Adding singularity manifest
This manifest tells our oscar cronjob which docker hub repositories to check
for updates. The file suffix is very important! The bnc website picks up card
data from files that end in `yaml`. Since we don't want the manifest to be
rendered as a card, we give it the sufix `yml`. File suffix doesn't matter to
the oscar cronjobs.